portfolio-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
pydantic
pydantic-settings
python-dotenv
psycopg2-binary